app install fix
