+++ packages/app-store/cal-ai/config.json
@@ -2,7 +2,7 @@
   "/*": "Don't modify slug - If required, do it using cli edit command",
   "name": "Cal AI",
   "slug": "cal-ai",
-  "type": "cal-ai_messaging",
+  "type": "cal-ai_automation",
   "logo": "icon.svg",
   "url": "https://cal.com/ai",
   "variant": "automation",
